# Performance improvement, typically 25%, by splitting internal 'is' function