@isaacs 2017-02-08
Landed on master and pushed to npm.  Thanks!