@lovell 2017-02-13
It looks like I underestimated the possible performance gains this change can bring, especially for very short XML strings.

By way of example, the node-expat [25 byte](https://github.com/astro/node-expat/blob/d5f8931b880b93080b156b8e569095278bcbdafe/benchmark.js#L61) [benchmark](https://github.com/astro/node-expat#benchmark) demonstrates a 2x improvement.

sax v1.2.1:
```
sax x 60,837 ops/sec ±1.54% (83 runs sampled)
node-expat x 440,272 ops/sec ±2.04% (83 runs sampled)
```
sax v1.2.2:
```
sax x 133,001 ops/sec ±1.59% (78 runs sampled)
node-expat x 440,120 ops/sec ±1.43% (85 runs sampled)
```
...so thank you for merging!